image-resizer-bot/requirements.txt
opencv-python
mysql
requests